// examples/typescript-react/src/schemas/api.ts
export const addressSchema = z.object({
  city: z.string(),
  country: z.string().regex(/^[A-Z]{2}$/),
  line1: z.string(),
  line2: z.string().optional(),
  postalCode: z.string(),
  state: z.string().optional(),
});
// ...
export const createOrderRequestSchema = z.object({
  billingAddress: z.lazy(() => addressSchema).optional(),
  notes: z.string().max(500).optional(),
  shippingAddress: z.lazy(() => addressSchema),
});